Refactor and add hot chats
